moved create onion to seprate file
