+++ .gitignore
@@ -5,3 +5,4 @@
 /**/__pycache__
 .DS_Store
 server/private_key
+private_key
